+++ src/BV/AABB.cpp
@@ -36,6 +36,7 @@
 /** \author Jia Pan */
 
 #include <hpp/fcl/BV/AABB.h>
+#include <hpp/fcl/shape/geometric_shapes.h>
 
 #include <limits>
 #include <hpp/fcl/collision_data.h>
@@ -142,6 +143,53 @@ bool overlap(const Matrix3f& R0, const Vec3f& T0, const AABB& b1,
              FCL_REAL& sqrDistLowerBound) {
   AABB bb1(translate(rotate(b1, R0), T0));
   return bb1.overlap(b2, request, sqrDistLowerBound);
+}
+
+bool AABB::overlap(const Plane& p) const {
+  // Convert AABB to a (box, transform) representation and compute the support
+  // points in the directions normal and -normal.
+  // If both points lie on different sides of the plane, there is an overlap
+  // between the AABB and the plane. Otherwise, there is no overlap.
+  const Vec3f halfside = (this->max_ - this->min_) / 2;
+  const Vec3f center = (this->max_ + this->min_) / 2;
+
+  const Vec3f support1 = (p.n.array() > 0).select(halfside, -halfside) + center;
+  const Vec3f support2 =
+      ((-p.n).array() > 0).select(halfside, -halfside) + center;
+
+  const FCL_REAL dist1 = p.n.dot(support1) - p.d;
+  const FCL_REAL dist2 = p.n.dot(support2) - p.d;
+  const int sign1 = (dist1 > 0) ? 1 : -1;
+  const int sign2 = (dist2 > 0) ? 1 : -1;
+
+  if (p.getSweptSphereRadius() > 0) {
+    if (sign1 != sign2) {
+      // Supports are on different sides of the plane. There is an overlap.
+      return true;
+    }
+    // Both supports are on the same side of the plane.
+    // We now need to check if they are on the same side of the plane inflated
+    // by the swept-sphere radius.
+    const FCL_REAL ssr_dist1 = std::abs(dist1) - p.getSweptSphereRadius();
+    const FCL_REAL ssr_dist2 = std::abs(dist2) - p.getSweptSphereRadius();
+    const int ssr_sign1 = (ssr_dist1 > 0) ? 1 : -1;
+    const int ssr_sign2 = (ssr_dist2 > 0) ? 1 : -1;
+    return ssr_sign1 != ssr_sign2;
+  }
+
+  return (sign1 != sign2);
+}
+
+bool AABB::overlap(const Halfspace& hs) const {
+  // Convert AABB to a (box, transform) representation and compute the support
+  // points in the direction -normal.
+  // If the support is below the plane defined by the halfspace, there is an
+  // overlap between the AABB and the halfspace. Otherwise, there is no
+  // overlap.
+  Vec3f halfside = (this->max_ - this->min_) / 2;
+  Vec3f center = (this->max_ + this->min_) / 2;
+  Vec3f support = ((-hs.n).array() > 0).select(halfside, -halfside) + center;
+  return (hs.signedDistance(support) < 0);
 }
 
 }  // namespace fcl